==> OSF Materials/03_Results.Rmd <==
---
title: "Results"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())
#install.packages("pacman")
pacman::p_load(tidyverse, readxl, ggbeeswarm, rptR, lme4, ggpubr, ggforce, magrittr, MuMIn)

###included colours are colourblind-friendly###
load("02_Run Models_Workspace.RData")
load("./Data/pilot_data.Rdata")

# model for "Do Zebrafish Show Appropriate Responses in an Aversive Learning Assay?"
learn_mod <- data.frame(coef(summary(learn_model_e1)))
learn_mod$coef_name <- rownames(learn_mod)

# models for "Do Zebrafish Display Colour Preferences and is Preference Repeatable?"    
grey_pref
green_pref
red_pref
orange_pref

# models for "Is Aversive Learning Repeatable Across and Within Conditions?"   
rpt_e1
rpt_allgb
rpt_allbg

# models for "Sex Differences in Repeatability"   
rpt_e1_male
rpt_e1_female
rpt_allbg_male
rpt_allbg_female
```

# Supplementary Tables   

## Supplementary Table 1   
```{r}
t1 <- read_excel("Results Tables.xlsx", sheet = 1, skip = 2)
t1[,2:4] <- apply(t1[,2:4], 2, function(x) as.numeric(x))

# estimate and p-values
t1[1:5,c(2,4)] <- round(learn_mod[,c("Estimate","Pr...t..")],3)

# confidence intervals
CIs <- round(data.frame(confint(learn_model_e1)),2)
t1$CI[1:5] <- paste0("'",CIs$X2.5..[4:8], " - ", CIs$X97.5..[4:8])

# variance components
varcor <- data.frame(summary(learn_model_e1)$varcor)
t1$Estimates[7:9] <- round(varcor$vcov[c(3,1,2)],2)
t1$Estimates[t1$Predictors=="ICC"] <- round(varcor$vcov[varcor$grp=="fishID"]/sum(varcor$vcov),2)

# sample sizes   
t1$Estimates[11:12] <- summary(learn_model_e1)$ngrps[c(2,1)]
t1$Estimates[13] <- length(learn_model_e1@resp$y)

# R2
R2 <- as.numeric(round(MuMIn::r.squaredGLMM(learn_model_e1),3))
t1$Estimates[14] <- paste0(R2[1],"/",R2[2])

#write.csv(t1,file="SupplementaryTable1.csv",row.names=F,na="")
```

## Supplementary Table 2    
```{r}
t2 <- read_excel("Results Tables.xlsx", sheet = 2, skip = 1)

# models included in this table
mod_list <- list(rpt_e1,
                 rpt_e1_male,
                 rpt_e1_female,
                 rpt_allgb,
                 rpt_allgb_male,
                 rpt_allgb_female,
                 rpt_allbg,
                 rpt_allbg_male,
                 rpt_allbg_female,
                 rpt_first,
                 rpt_first_male,
                 rpt_first_female,
                 rpt_second,
                 rpt_second_male,
                 rpt_second_female,
                 rpt_third,
                 rpt_third_male,
                 rpt_third_female,
                 rpt_allgbbg,
                 rpt_allgbbg_male,
                 rpt_allgbbg_female)

# position of rows with data in table
pos <- (!is.na(t2$`Repeatability (R)`))

t2$`Repeatability (R)`[pos] <- sapply(mod_list, function(x) round(x$R,3)) %>% unlist 
t2$`95% Confidence Interval`[pos] <- sapply(mod_list, function(x) paste0(round(x$CI_emp[1],3), " - ", round(x$CI_emp[2],3))) %>% unlist

#write.csv(t2,file="SupplementaryTable2.csv",row.names=F,na="")
```

## Supplementary Table 3    
```{r}
t3 <- read_excel("Results Tables.xlsx", sheet = 3, skip = 1)

# models included in this table
mod_list <- list(red_pref,
                 red_pref_male,
                 red_pref_female,
                 green_pref,
                 green_pref_male,
                 green_pref_female,
                 grey_pref,
                 check_pref_male,
                 check_pref_female,
                 orange_pref,
                 orange_pref_male,
                 orange_pref_female)

# position of rows with data in table
pos <- (!is.na(t3$`Repeatability (R)`))

t3$`Repeatability (R)`[pos] <- sapply(mod_list, function(x) round(x$R,3)) %>% unlist 
t3$`95% Confidence Interval`[pos] <- sapply(mod_list, function(x) paste0(round(x$CI_emp[1],3), " - ", round(x$CI_emp[2],3))) %>% unlist

#write.csv(t3,file="SupplementaryTable3.csv",row.names=F,na="")
```

# Figures   
## Tidying Data for Figures
```{r, Wrangling}
rpt_outputs$conditions <- as.factor(rpt_outputs$conditions)

rpt_outputs %<>% mutate(CS = case_when (grepl("Blue_Green", conditions) ~ "Blue/Green",
                                        grepl("Blue_Green", conditions) ~ "Blue/Green",
                                        grepl("Green_Blue", conditions) ~ "Green/Blue",
                                        grepl("Green_Blue", conditions) ~ "Green/Blue",
                                        grepl("First", conditions) ~ "Green/Blue & Blue/Green",
                                        grepl("Second", conditions) ~ "Green/Blue & Blue/Green",
                                        grepl("Third", conditions) ~ "Green/Blue & Blue/Green",
                                        grepl("All", conditions) ~ "Green/Blue & Blue/Green",
                                        grepl("Across", conditions) ~ "Across Conditions"))

rpt_outputs$CS = factor(rpt_outputs$CS, levels = c('Across Conditions','Green/Blue','Blue/Green','Green/Blue & Blue/Green')) #ordering factors correctly for visualisation
rpt_outputs$conditions = factor(rpt_outputs$conditions, levels = c("Across","Green_Blue","Blue_Green", "All", "Third", "Second", "First"))

rpt_outputs %>% slice(1:3) -> new_outputs #remove measurement set data
```

## Figure 2   
```{r}
# colours
pref_cols <- c("Grey" = "#999999","Green" = "#009E73", "Red" = "#D55E00", "Orange" = "#E69F00")

# ggplot theme
tm <- theme(panel.background = element_blank(),
            panel.grid = element_line(colour = "grey", size = 0.2),
            axis.line = element_line(colour = "black"),
            axis.ticks = element_line(colour = "black"),
            axis.title.x = element_blank(),
            axis.text.x = element_text(size = 11, colour = "black"),
            axis.text.y = element_text(size = 9, colour = "black"),
            axis.title.y = element_text(size = 11, vjust = 3),
            plot.tag = element_text(size = 13, face = "bold"),
            plot.title = element_text(size = 13, hjust = 0.5))

col_pref$stim <- factor(col_pref$stim, levels = c("Grey", "Green", "Red", "Orange"))
col_pref %>% group_by(stim) %>% summarise(mu = mean(pref/60),
                                          sigma = sd(pref/60),
                                          n = n(),
                                          se = sigma/sqrt(n)) -> pref_stats

  ggplot() + tm + 
  geom_hline(yintercept = 0.5, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_quasirandom(data = col_pref, aes(stim, pref/60, col = stim),
                   dodge.width = 0.4, alpha = 0.5,
                   show.legend = F) +
  geom_errorbar(data = pref_stats, aes(x = stim, 
                                 group = stim,
                                 ymin = mu - sigma,
                                 ymax = mu + sigma),
                width = 0.3,
                size = 0.5) +
    geom_point(data = pref_stats, aes(x = stim, 
                                 group = stim,
                                 y = mu),
               size = 3) +
  scale_color_manual(values = pref_cols) +
  scale_y_continuous(breaks = seq(0,1,0.1), lim = c(0,1)) +
  labs(x = "Colour", y = "proportion of time spent in colour", title = "", tag = "(A)") +
  scale_x_discrete(labels = c("Grey > Check", "Green > Blue", "Red > Yellow", "Orange > Magenta")) -> A
  
 
  title <- ggdraw() +
  draw_label(
    "Baseline Colour Preferences",
    fontface = 'bold',
    x = 0.5,
    hjust = 0.5
  ) +
  theme(
    # add margin on the left of the drawing canvas,
    # so title is aligned with left edge of first plot
    plot.margin = margin(0, 0, 0, 7)
  )
top_panel <- plot_grid(title, A, ncol = 1, rel_heights = c(0.1,1))
  

female.colour = "#730091"
male.colour = "#088701"

# e1 %>% group_by(sex) %>% summarise(mu = mean(difference/60),
#                                    sigma = sd(difference/60),
#                                    min = min(difference/60),
#                                    max = max(difference/60),
#                                    n = n(),
#                                    se = sigma/sqrt(n)) -> learn_stats1
# 
# e2 %>% group_by(sex) %>% summarise(mu = mean(difference/60),
#                                    sigma = sd(difference/60),
#                                    min = min(difference/60),
#                                    max = max(difference/60),
#                                    n = n(),
#                                    se = sigma/sqrt(n)) -> learn_stats2
# 
# range_learn_stats <- range(c(learn_stats1$min,
#                              learn_stats1$max,
#                              learn_stats2$min,
#                              learn_stats2$max))
# 
# ggplot() + tm + 
#   geom_hline(yintercept = 0, linetype = "dashed", colour = "grey40", size = 0.5) +
#   geom_quasirandom(data = e1, aes(sex, difference/60, col = sex),
#                    dodge.width = 0.4, alpha = 0.5,
#                    show.legend = F) +
#   geom_errorbar(data = learn_stats1, aes(x = sex, 
#                                  group = sex,
#                                  ymin = mu - sigma,
#                                  ymax = mu + sigma),
#                 width = 0.3,
#                 size = 0.5) +
#     geom_point(data = learn_stats1, aes(x = sex, 
#                                  group = sex,
#                                  y = mu),
#                size = 3) +
#   scale_color_manual(values = c(female.colour, male.colour)) +
#   scale_y_continuous(breaks = seq(-1,1,0.1), lim = range_learn_stats) +
#   labs(x = "Colour", y = "proportional difference to baseline time", title = "across conditions", tag = "(B)") -> B
# 
# 
# 
# ggplot() + tm + theme(axis.text.y = element_blank(),
#                       axis.title.y = element_blank(),
#                       axis.ticks.y = element_blank(),
#                       axis.line.y = element_blank()) +
#   geom_hline(yintercept = 0, linetype = "dashed", colour = "grey40", size = 0.5) +
#   geom_quasirandom(data = e2, aes(sex, difference/60, col = sex),
#                    dodge.width = 0.4, alpha = 0.5,
#                    show.legend = F) +
#   geom_errorbar(data = learn_stats2, aes(x = sex, 
#                                  group = sex,
#                                  ymin = mu - sigma,
#                                  ymax = mu + sigma),
#                 width = 0.3,
#                 size = 0.5) +
#     geom_point(data = learn_stats2, aes(x = sex, 
#                                  group = sex,
#                                  y = mu),
#                size = 3) +
#   scale_color_manual(values = c(female.colour, male.colour)) +
#   scale_y_continuous(breaks = seq(-1,1,0.1), lim = range_learn_stats) +
#   labs(x = "Colour", y = "proportional difference to baseline time", title = "within conditions", tag = "(C)") -> C
# 
# bottom <- plot_grid(B, C, rel_widths = c(0.53,0.47))
#bottom

# putting across and within-conditions data together
e_both <- rbind(e1[,c("sex", "difference")],
                e2[,c("sex", "difference")])

e_both %>% group_by(sex) %>% summarise(mu = mean(difference/60),
                                   sigma = sd(difference/60),
                                   min = min(difference/60),
                                   max = max(difference/60),
                                   n = n(),
                                   se = sigma/sqrt(n)) -> learn_stats

range_learn_stats <- range(c(learn_stats$min,
                             learn_stats$max))

ggplot() + tm +
  geom_hline(yintercept = 0, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_quasirandom(data = e_both, aes(sex, difference/60, col = sex),
                   dodge.width = 0.4, alpha = 0.5,
                   show.legend = F) +
  geom_errorbar(data = learn_stats, aes(x = sex,
                                 group = sex,
                                 ymin = mu - sigma,
                                 ymax = mu + sigma),
                width = 0.3,
                size = 0.5) +
    geom_point(data = learn_stats, aes(x = sex,
                                 group = sex,
                                 y = mu),
               size = 3) +
  scale_color_manual(values = c(female.colour, male.colour)) +
  scale_y_continuous(breaks = seq(-1,1,0.1), lim = range_learn_stats) +
  labs(x = "Colour", y = "proportional difference to baseline time", tag = "(B)") -> bottom

ggdraw(bottom) +
  draw_line(x = c(0.545, 0.545),
    y = c(0.42, 0.85),
    colour = "#ebebeb",
    size = 6.5,
    linejoin = "mitre",
    arrow = arrow(type="closed")) +
  draw_line(x = c(0.545, 0.545),
    y = c(0.35, 0.1),
    colour = "#ebebeb",
    size = 6.5,
    linejoin = "mitre",
    arrow = arrow(type="closed")) +
  draw_label("aversion", angle = 90, colour = "black", size = 11,
             x = 0.544, y = 0.52) +
  draw_label("preference", angle = 90, colour = "black", size = 11,
             x = 0.544, y = 0.24) -> bottom_panel

title <- ggdraw() +
  draw_label(
    "Aversive Learning",
    fontface = 'bold',
    x = 0.5,
    hjust = 0.5
  ) +
  theme(
    # add margin on the left of the drawing canvas,
    # so title is aligned with left edge of first plot
    plot.margin = margin(10, 0, 0, 7)
  )

bottom <- plot_grid(title, bottom_panel, ncol = 1, rel_heights = c(0.1,1))

Fig2 <- plot_grid(top_panel, bottom, nrow = 2,
                  rel_heights = c(0.48, 0.52))
ggsave(filename = "./Figures/Figure 2.pdf",
       plot = Fig2,
       height = 8, width = 9,
       device = cairo_pdf)
```

## Alternative Figure 2
```{r, Results Figure 2}
pref_cols <- c("Grey" = "#999999","Green" = "#009E73", "Red" = "#D55E00", "Orange" = "#E69F00")

sex_cols1 <- c("male" = "#56B4E9", "female" = "#CC79A7")

fig2A <-  e1 %>% 
  ggplot(aes(sex, difference,  col = sex)) + 
  geom_hline(yintercept = 0, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_hline(yintercept = 2.695262, colour = "#CC79A7", size = 1.5) +
  geom_hline(yintercept = 5.063168, colour = "#56B4E9", size = 1.5) +
  geom_quasirandom(dodge.width = 0.6, alpha = 0.5) + 
  theme(
    legend.position = c(.95, .95),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6)) +
  labs(col = "Sex") +
  labs(x = "Sex", y = "Association Response", title = "Aversive Learning in Zebrafish") +
  scale_color_manual(values = sex_cols1) +
  theme(legend.key = element_rect(fill = "white", colour = "grey90")) +
  theme(axis.text.x = element_text(size = 11))

fig2B <- col_pref %>% 
  ggplot(aes(stim, pref, col = stim)) + 
  geom_hline(yintercept = 30, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_quasirandom(dodge.width = 0.4, alpha = 0.5) +
  scale_color_manual(values = pref_cols) +
  labs(x = "Colour", y = "Relative Colour Preference", title = "Colour Preference During the Baseline Period") +
  theme(legend.position = "none") +
  scale_x_discrete(name = "Stimulus",labels = c("Grey\nCheck", "Green\nBlue", "Red\nYellow", "Orange\nMagenta")) +
  theme(axis.text.x = element_text(size = 11))

Figure_2 <- ggarrange(fig2A, fig2B,
                    labels = c("A", "B"),
                    ncol = 1, nrow = 2)

Figure_2
```

## Alternative Figure 2.1
```{r, Results Figure 2}
#FIGURE IN PROGRESS
##mean lines are wrong
pref_cols <- c("Grey" = "#999999","Green" = "#009E73", "Red" = "#D55E00", "Orange" = "#E69F00")

sex_cols1 <- c("male" = "#CC79A7", "female" = "#56B4E9")

bothexp %<>% mutate(experiment = case_when (grepl("Experiment 1", filepath) ~ "Experiment 1",
                                            grepl("MOBASELINE", filepath) ~ "Experiment 1",
                                           grepl("Experiment 2", filepath) ~ "Experiment 2"))

bothexp$expsex <- as.factor(paste(bothexp$experiment, bothexp$sex, sep = " "))

fig2A.1 <- e1 %>% 
  ggplot(aes(sex, difference,  col = sex)) + 
  geom_hline(yintercept = 0, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_hline(yintercept = 2.9454, colour = "#CC79A7", size = 1.5) +
  geom_hline(yintercept = 6.0191, colour = "#56B4E9", size = 1.5) +
  geom_quasirandom(dodge.width = 0.6, alpha = 0.5) + 
  theme(
    legend.position = c(.95, .95),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6)) +
  labs(col = "Sex") +
  labs(x = "Sex", y = "Association Response", title = "Aversive Learning in Experiment 1") +
  scale_color_manual(values = sex_cols1) +
  theme(legend.key = element_rect(fill = "white", colour = "grey90")) +
  ylim(-22,41) +
  theme(axis.text.x = element_text(size = 11))

fig2B.1 <- e2 %>% 
  ggplot(aes(sex, difference,  col = sex)) + 
  geom_hline(yintercept = 0, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_hline(yintercept = 6.88287, colour = "#CC79A7", size = 1.5) +
  geom_hline(yintercept = 7.5871, colour = "#56B4E9", size = 1.5) +
  geom_quasirandom(dodge.width = 0.6, alpha = 0.5) + 
  theme(
    legend.position = c(.95, .95),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(6, 6, 6, 6)) +
  labs(col = "Sex") +
  labs(x = "Sex", y = "Association Response", title = "Aversive Learning in Experiment 2") +
  scale_color_manual(values = sex_cols1) +
  theme(legend.key = element_rect(fill = "white", colour = "grey90")) +
  ylim(-22,41) +
  theme(axis.text.x = element_text(size = 11))

fig2C.1 <- col_pref %>% 
  ggplot(aes(stim, pref, col = stim)) + 
  geom_hline(yintercept = 30, linetype = "dashed", colour = "grey40", size = 0.5) +
  geom_quasirandom(dodge.width = 0.4, alpha = 0.5) +
  scale_color_manual(values = pref_cols) +
  labs(x = "Colour", y = "Relative Colour Preference", title = "Colour Preference During the Baseline Period in Experiment 1") +
  theme(legend.position = "none") +
  scale_x_discrete(name = "Stimulus",labels = c("Grey\nCheck", "Green\nBlue", "Red\nYellow", "Orange\nMagenta")) +
  theme(axis.text.x = element_text(size = 11))

Figure_2.1 <- ggarrange(fig2C.1,
                        ggarrange(fig2A.1, fig2B.1, ncol = 2, labels = c("B", "C")),
                        nrow = 2,
                        labels = "A"
)

Figure_2.1
```
## Figure 3    
```{r}
# colours
pref_cols <- c("Grey" = "#999999","Green" = "#009E73", "Red" = "#D55E00", "Orange" = "#E69F00")
new_cols <- c("Blue_Green" = "#56B4E9", "Green_Blue" = "#009E73", "Across" = "#000000")
# ggplot theme
tm <- theme(panel.background = element_blank(),
            panel.grid = element_line(colour = "grey", size = 0.2),
            axis.line = element_line(colour = "black"),
            axis.ticks = element_line(colour = "black"),
            axis.title.y = element_blank(),
            axis.text.y = element_text(size = 11, colour = "black"),
            axis.text.x = element_text(size = 9, colour = "black", margin = margin(0,0,0.1,0,"cm")),
            axis.title.x = element_text(size = 10, face = "italic"),
            plot.tag = element_text(size = 13, face = "bold"),
            plot.title = element_text(size = 13, hjust = 0.5))

# ordering factor to match colours
pref_outputs$conditions <- factor(pref_outputs$conditions, levels = c("Grey", "Green", "Red", "Orange"))
new_outputs$conditions <- factor(new_outputs$conditions, levels = c("Blue_Green", "Green_Blue", "Across"))

ggplot(data = pref_outputs, aes(x = conditions, y = R, colour = conditions)) + tm +
  geom_pointrange(aes(y = R, ymin = lowCI, ymax = highCI), size = 1, show.legend = F, fatten = 5) +
  scale_x_discrete(labels = c("Grey > Check", "Green > Blue", "Red > Yellow", "Orange > Magenta")) +
  scale_y_continuous(n.breaks = 6, expand = c(0,0), lim = c(0,max(pref_outputs$highCI))) +
  labs(x = "Colour", y = "enhanced repeatability", title = "Repeatabiltiy of Colour Preference",
       tag = "(A)") +
  scale_color_manual(values = pref_cols) +
  coord_flip() -> A


ggplot(data = new_outputs, aes(x = conditions, y = R, colour = conditions)) + tm +
  geom_pointrange(aes(y = R, ymin = lowCI, ymax = highCI), size = 1, show.legend = F, fatten = 5) +
  scale_x_discrete(labels = c("Blue aversion (within pair)", "Green aversion (within pair)", "Across Pairs")) +
  scale_y_continuous(n.breaks = 6, expand = c(0,0), lim = c(0,max(pref_outputs$highCI))) +
  labs(x = "Colour", y = "enhanced repeatability", title = "Repeatabiltiy of Aversive Learning",
       tag = "(B)") +
  scale_color_manual(values = new_cols) +
  coord_flip() -> B

Fig3 <- plot_grid(A,B,nrow = 2)
ggsave(filename = "./Figures/Figure 3.pdf",
       plot = Fig3,
       height = 6, width = 8,
       device = cairo_pdf)
```

## Alternative Figure 3
```{r, Results Figure 3}
##forest plot showing r estimate and 95% CI, conditions grouped according to CS+
new_cols <- c("Across" = "#000000","Green_Blue" = "#009E73", "Blue_Green" = "#56B4E9")

fig3A <- ggplot(data = pref_outputs, aes(x = conditions, y = R, colour = conditions)) +
  geom_errorbar(aes(ymin = lowCI, ymax = highCI), width = 0.15, size = 0.6, position = position_dodge(0.3)) +
  geom_point(aes(x = conditions, y = R), position = position_dodge(0.3), size = 1.5) +
  geom_hline(yintercept = 0, lty = "dotted") +
  labs(x = "Colour", y = "Repeatability (R)", title = "Repeatabiltiy of Colour Preference") +
  facet_col(vars(conditions), scales = "free_y", space = "free") +
  scale_color_manual(values = pref_cols) +
  ylim(0, 0.63) +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank()) +
  theme(legend.position = "none") +
  theme(axis.text.x = element_text(size = 11)) +
  theme(strip.text.x = element_text(size = 11)) +
  coord_flip()

fig3B <- ggplot(data = new_outputs, aes(x = conditions, y = R, colour = conditions)) +
  geom_errorbar(aes(ymin = lowCI, ymax = highCI), width = 0.15, size = 0.6, position = position_dodge(0.3)) +
  geom_point(aes(x = conditions, y = R), position = position_dodge(0.3), size = 1.5) +
  geom_hline(yintercept = 0, lty = "dotted") +
  labs(x = "Condition", y = "Repeatability (R)", title = "Repeatabiltiy of Aversive Learning") +
  facet_col(vars(CS), scales = "free_y", space = "free") + #facet_col produces a better forest plot than facet_wrap
  scale_colour_manual(values = new_cols) +
  ylim(0, 0.63) +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank()) +
  theme(axis.text.x = element_text(size = 11)) +
  theme(strip.text.x = element_text(size = 11)) +
  theme(legend.position = "none") +
  coord_flip()

Figure_3 <- ggarrange(fig3A, fig3B,
                    labels = c("A", "B"),
                    ncol = 1, nrow = 2,
                    heights = c(1.1,0.8))
Figure_3
```

# Secondary Visualisation
## Exp and Sex differences
```{r, Beeswarm plots of sex & exp in different conditions}
#Across Conditions sex differences
ggplot(e1, aes(sex, difference, color=sex)) + 
  geom_quasirandom(dodge.width=0.4, alpha=0.7)

#Across Conditions exp differences
ggplot(e1, aes(sex, difference, color=sex)) + 
  geom_quasirandom(dodge.width=0.4, alpha=0.7) +
  facet_wrap(~exp)

#Both Experiment sex differences
ggplot(allexp, aes(sex, difference, color=sex)) + 
  geom_quasirandom(dodge.width=0.4, alpha=0.7)

#Green/Blue sex differences
ggplot(allgb, aes(sex, difference, color=sex)) + 
  geom_quasirandom(dodge.width=0.2, alpha=0.7)

#Blue/Green sex differences
ggplot(allbg, aes(sex, difference, color=sex)) + 
  geom_quasirandom(dodge.width=0.2, alpha=0.7)
```

# Supplementary Figures  
## Supplementary Figure 3 
```{r forest plot with all conditions and sex groupings}
#renaming values to be put in figures
rpt_outputs_all$conditions <- gsub("\\_M", " M", rpt_outputs_all$conditions)
rpt_outputs_all$conditions <- gsub("\\_F", " F", rpt_outputs_all$conditions)
rpt_outputs_all$conditions <- gsub("\\_", " ", rpt_outputs_all$conditions)

rpt_outputs_all[1, 1] <- "Across"
rpt_outputs_all[2, 1] <- "Green/Blue"
rpt_outputs_all[3, 1] <- "Blue/Green"
rpt_outputs_all[4, 1] <- "First"
rpt_outputs_all[5, 1] <- "Second"
rpt_outputs_all[6, 1] <- "Third"
rpt_outputs_all[7, 1] <- "All"
rpt_outputs_all[8, 1] <- "Across Male"
rpt_outputs_all[9, 1] <- "Across Female"
rpt_outputs_all[10, 1] <- "Green/Blue Male"
rpt_outputs_all[11, 1] <- "Green/Blue Female"
rpt_outputs_all[12, 1] <- "Blue/Green Male"
rpt_outputs_all[13, 1] <- "Blue/Green Female"
rpt_outputs_all[14, 1] <- "First Male"
rpt_outputs_all[15, 1] <- "First Female"
rpt_outputs_all[16, 1] <- "Second Male"
rpt_outputs_all[17, 1] <- "Second Female"
rpt_outputs_all[18, 1] <- "Third Male"
rpt_outputs_all[19, 1] <- "Third Female"
rpt_outputs_all[20, 1] <- "All Male"
rpt_outputs_all[21, 1] <- "All Female"

rpt_outputs_all %<>% mutate(Sex = case_when (grepl("Male", conditions) ~ "Male",
                                             grepl("Female", conditions) ~ "Female",
                                             grepl("Green/Blue", conditions) ~ "Mixed",
                                             grepl("Blue/Green", conditions) ~ "Mixed",
                                             grepl("First", conditions) ~ "Mixed",
                                             grepl("Second", conditions) ~ "Mixed",
                                             grepl("Third", conditions) ~ "Mixed",
                                             grepl("All", conditions) ~ "Mixed",
                                             grepl("Across", conditions) ~ "Mixed")) %>% 
                     mutate(CS = case_when (grepl("Blue/", conditions) ~ "Blue",
                                            grepl("Green/", conditions) ~ "Green",
                                            grepl("First", conditions) ~ "Both",
                                            grepl("Second", conditions) ~ "Both",
                                            grepl("Third", conditions) ~ "Both",
                                            grepl("All", conditions) ~ "Both",
                                            grepl("Across", conditions) ~ "All"))

rpt_outputs_all[sapply(rpt_outputs_all, is.character)] <- lapply(rpt_outputs_all[sapply(rpt_outputs_all, is.character)], as.factor) #turning chr to factors

sex_cols <- c("Female" = "#730091","Male" = "#088701", "Mixed" = "#999999")

rpt_outputs_all$CS = factor(rpt_outputs_all$CS, levels = c('All','Green','Blue','Both')) #ordering factors correctly for visualisation

rpt_outputs_all$conditions = factor(rpt_outputs_all$conditions, levels = c("Across","Across Male","Across Female","Green/Blue", "Green/Blue Male", "Green/Blue Female", "Blue/Green", "Blue/Green Male", "Blue/Green Female","First",  "First Male", "First Female", "Second", "Second Male", "Second Female", "Third", "Third Male", "Third Female", "All", "All Male", "All Female"))

##forest plot showing r estimate and 95% CI, conditions grouped according to CS+
Figure_S3 <- ggplot(data = rpt_outputs_all, aes(x = conditions, y = R, colour = Sex)) +
  geom_errorbar(aes(ymin = lowCI, ymax = highCI), width = 0.05, position = position_dodge(0.3)) +
  geom_point(aes(x = conditions, y = R), position = position_dodge(0.3))+
  geom_hline(yintercept = 0, lty = "dotted") +
  labs(x = "Condition", y = "Repeatability (R)") +
  facet_col(vars(CS), scales = "free_y", space = "free") + #facet_col produces a better forest plot than facet_wrap
  scale_colour_manual(values = sex_cols) +
  coord_flip()

Figure_S3
```
## Supplementary Figure 4
```{r, Constrast Analysis Plot}
sex_cols <- c("Female" = "#730091","Male" = "#088701", "Mixed" = "#999999")

Figure_S4 <- ggplot(data = df_contrast_learn, aes(x = Distribution, y = R, colour = Sex)) +
  geom_errorbar(aes(ymin = lowCI, ymax = highCI), width = 0.1, position = position_dodge(0.3), size = 1) +
  geom_point(aes(x = Distribution, y = R), position = position_dodge(0.3), size = 3) +
  geom_hline(yintercept = 0, lty = "dotted") +
  labs(x = "Distribution", y = "Repeatability (R)") +
  scale_colour_manual(values = sex_cols) +
  theme(legend.position="none") +
  coord_flip()

Figure_S4
```

## Supplenmetary Figure 5  
```{r}
volt_data %>% filter(test == "CHECK") %>% group_by(test, voltage) %>% summarise(mean = mean(difference), sd = sd(difference), se = sd(difference)/sqrt(n())) -> volt_sum

Figure_S5 <- ggplot(volt_sum, aes(voltage, mean)) +
  geom_bar(position=position_dodge(), stat= "identity", width = 0.7, colour="black") +
  geom_errorbar(aes(ymin=mean - se, ymax = mean + se), width = 0.4, colour = "black", position=position_dodge(0.9)) +
  xlab("Unconditioned Stimulus Voltage") +
  ylab("Association Response (seconds per minute) ± SD") +
  ylim(0,15)
```


## Supplementary Figure 6
```{r}
e1_data %>% filter(zone == "CS") %>% group_by(exp, unit) %>% 
  summarise(mean=mean(time_min), sd=sd(time_min), se=sd(time_min)/sqrt(n())) -> e1_sum
e1_sum %>% group_by(unit) %>% 
  summarise(mean=mean(mean), sd=sd(sd), se=sd(se)/sqrt(n())) -> e1_sum_all

probe_draw_all <- function(dat, title = "") {
  dat <- filter(dat, unit %in% c("PROBE1", "PROBE2", "PROBE3", "PROBE4", "PROBE5"))
  gplot <- ggplot(dat, aes(unit, mean)) +
  geom_bar(position=position_dodge(), stat= "identity", width = 0.7, colour="black") +
  geom_errorbar(aes(ymin=mean - sd, ymax = mean + sd), width = 0.4, colour = "black", position=position_dodge(0.9)) +
  labs(x = "Minute Increments in Probe Period", y = "Time Spent in CS+ During Probe (seconds per minute) ± SD", title = title) +
  ylim(0,45)

  print(gplot)
}

probe_draw_all(e1_sum_all, title =  "All Conditions") -> Figure_S6
```

## Supplementary Figure 7
```{r}
e1_data %>% filter(zone == "CS") %>% group_by(exp, unit) %>% 
  summarise(mean=mean(time_min), sd=sd(time_min), se=sd(time_min)/sqrt(n())) -> e1_sum

probe_draw <- function(dat, condition = "", title = "") {
  dat <- filter(dat, exp == condition)
  dat <- filter(dat, unit %in% c("PROBE1", "PROBE2", "PROBE3", "PROBE4", "PROBE5"))
  gplot <- ggplot(dat, aes(unit, mean)) +
  geom_bar(position=position_dodge(), stat= "identity", width = 0.7, colour="black") +
  geom_errorbar(aes(ymin=mean - se, ymax = mean + se), width = 0.4, colour = "black", position=position_dodge(0.9)) +
  labs(x = "Minute  Increments in Probe Period", y = "CS+ avoidance ± SE", title = title) +
  ylim(0,45)
  
  print(gplot)
}

probe_draw(e1_sum, condition = "REDYEL", title = "Red/Yellow Condition") -> p1
probe_draw(e1_sum, condition = "YELRED", title = "Yellow/Red Condition") -> p2
probe_draw(e1_sum, condition = "GREBLU", title = "Green/Blue Condition") -> p3
probe_draw(e1_sum, condition = "BLUGRE", title = "Blue/Green Condition") -> p4
probe_draw(e1_sum, condition = "CHKGRY", title = "Check/Grey Condition") -> p5
probe_draw(e1_sum, condition = "GRYCHK", title = "Grey/Check Condition") -> p6
probe_draw(e1_sum, condition = "ORAMAG", title = "Orange/Magenta Condition") -> p7
probe_draw(e1_sum, condition = "MAGORA", title = "Magenta/Orange Condition") -> p8


Figure_S7 <- ggarrange(p1, p2, p2, p4, p5, p6, p7, p8,
                    ncol = 2, nrow = 4)

Figure_S7
```
